pyproject.toml
[project]
name = "rag-property-agent"
version = "0.1.0"
description = "Multi-Agent RAG System for Property Sales Assistance - MVP for Thesis"
authors = [{ name = "Developer", email = "dev@example.com" }]
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }

dependencies = [
    # Core AI Framework
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    
    # Vector Database
    "chromadb>=0.5.0",
    
    # Database
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "asyncpg>=0.29.0",
    
    # Cache & Session
    "redis>=5.0.0",
    
    # HTTP & API
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "httpx>=0.27.0",
    
    # Data Processing
    "pandas>=2.0.0",
    "numpy>=1.26.0",
    
    # Environment & Config
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    
    # Utilities
    "tenacity>=8.0.0",
    "structlog>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "black>=24.0.0",
    "ruff>=0.4.0",
    "mypy>=1.0.0",
]

analysis = [
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "scipy>=1.12.0",
    "jupyter>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
markers = [
    "integration: marks tests as integration tests (require API keys, databases)",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311"]
